Add authentication to mutations

+++ graphql-prisma/src/resolvers/Mutation.js
@@ -1,5 +1,6 @@
 import bcrypt from 'bcryptjs'
 import jwt from 'jsonwebtoken'
+import getUserId from '../utils/getUserId'
 
 const Mutation = {
     async createUser(parent, args, { prisma }, info) {
@@ -42,24 +43,29 @@ const Mutation = {
             user,
             token: jwt.sign({ userId: user.id }, 'thisisasecret')
         }
-
     },
-    deleteUser(parent, args, { prisma }, info) {
+    deleteUser(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+
         return prisma.mutation.deleteUser({
             where: {
-                id: args.id
+                id: userId
             }
         }, info)
     },
-    updateUser(parent, { id, data }, { prisma }, info) {
+    updateUser(parent, { data }, { prisma, request }, info) {
+        const userId = getUserId(request)
+
         return prisma.mutation.updateUser({
             where: {
-                id
+                id: userId
             },
             data
         }, info)
     },
-    createPost(parent, args, { prisma }, info) {
+    createPost(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+
         return prisma.mutation.createPost({
             data: {
                 title: args.data.title,
@@ -67,20 +73,46 @@ const Mutation = {
                 published: args.data.published,
                 author: {
                     connect: {
-                        id: args.data.author
+                        id: userId
                     }
                 }
             }
         }, info)
     },
-    deletePost(parent, args, { prisma }, info) {
+    async deletePost(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+
+        const postExists = await prisma.exists.Post({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+
+        if(!postExists) {
+            throw new Error("Unable to delete post")
+        }
+
         return prisma.mutation.deletePost({
             where: {
                 id: args.id
             }
         }, info)
     },
-    updatePost(parent, { id, data }, { prisma }, info) {
+    async updatePost(parent, { id, data }, { prisma, request }, info) {
+        const userId = getUserId(request)
+
+        const postExists = await prisma.exists.Post({
+            id,
+            author: {
+                id: userId
+            }
+        })
+
+        if(!postExists) {
+            throw new Error("Unable to update post")
+        }
+
         return prisma.mutation.updatePost({
             where: {
                 id
@@ -88,13 +120,15 @@ const Mutation = {
             data
         }, info)
     },
-    createComment(parent, { data }, { prisma }, info) {
+    createComment(parent, { data }, { prisma, request }, info) {
+        const userId = getUserId(request)
+
         return prisma.mutation.createComment({
             data: {
                 text: data.text,
                 author: {
                     connect: {
-                        id: data.author
+                        id: userId
                     }
                 },
                 post: {
@@ -105,14 +139,40 @@ const Mutation = {
             }
         }, info)
     },
-    deleteComment(parent, args, { prisma }, info) {
+    async deleteComment(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+
+        const commentExists = await prisma.exists.Comment({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+
+        if(!commentExists) {
+            throw new Error('Unable to delete comment')
+        }
+
         return prisma.mutation.deleteComment({
             where: {
                 id: args.id
             }
         }, info)
     },
-    updateComment(parent, { id, data }, { prisma }, info) {
+    async updateComment(parent, { id, data }, { prisma, request }, info) {
+        const userId = getUserId(request)
+
+        const commentExists = await prisma.exists.Comment({
+            id,
+            author: {
+                id: userId
+            }
+        })
+
+        if(!commentExists) {
+            throw new Error("Unable to update comment")
+        }
+
         return prisma.mutation.updateComment({
             where: {
                 id
